+++ src/core/compiler/generator.ts
@@ -1,9 +1,10 @@
 
-import { type Component, h, type VNode } from 'vue'
+import { h, type VNode } from 'vue'
 import { ITreeNode } from '@/core/compiler/types'
+import { IComponentInternal } from '@/core/types'
 
 
-export function generateVNodes(tree: ITreeNode, componentsMap: Record<string, Component>): VNode {
+export function generateVNodes(tree: ITreeNode, componentsMap: Record<string, IComponentInternal>): VNode {
   // Base case of recursion
   if (tree.children.length === 0) {
     switch (tree.category) {
@@ -16,7 +17,7 @@ export function generateVNodes(tree: ITreeNode, componentsMap: Record<string, Co
         if (!component) {
           throw new Error(`Component ${tree.tagName} not found`)
         }
-        return h(component, tree.properties)
+        return h(component.shallowRef, tree.properties)
       case 'root':
         return h('div', {class: "body--root"})
     }