+++ src/core/index.ts
@@ -1,3 +1,5 @@
 import { useCMS } from './composables/useCMS'
+import Editor from './components/editor/wrapper.vue'
+import ContentBody from './components/content/body.vue'
 
-export { useCMS }
+export { useCMS, Editor, ContentBody }