+++ src/playground/cmp-wit-prop.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+
+const props = defineProps<{display: string}>()
+</script>
+
+<template>
+  <div>
+    {{ props.display }}
+  </div>
+</template>
+
+