+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-### 0.0.5
+### 0.1.0
+- useCMS composable
+- Content Body view
+- Basic editor setup
 
+### 0.0.5
 - Testing approach for of modularization into `core` and `utils`